refactor: User에서 인증 관련 기능을 Auth로 빼서 관리
