+++ backend/src/main/java/endolphin/backend/global/google/dto/GoogleUserInfo.java
@@ -1,4 +1,4 @@
-package endolphin.backend.domain.user.dto;
+package endolphin.backend.global.google.dto;
 
 public record GoogleUserInfo(String sub, String name, String email, String picture) {
